DB Scripts and Player & team strength representation in floating point

See merge request courses/2020-fall/csci-5308/group6/csci5308!41

+++ src/main/java/com/IceHockeyLeague/LeagueManager/Player/IPlayer.java
@@ -17,7 +17,7 @@ public interface IPlayer {
     IPlayerStats getPlayerStats();
     void setPlayerStats(IPlayerStats playerStats);
 
-    double calculateStrength(IPlayerStats stats);
+    float calculateStrength(IPlayerStats stats);
 
     boolean isValid();
 }
